Revert grpc refactor; add minimal EPD hook
