+++ python/sglang/srt/grpc/grpc_request_manager.py
@@ -17,16 +17,12 @@ from typing import Any, AsyncGenerator, Dict, List, Optional, Union
 import grpc
 import zmq
 import zmq.asyncio
-from google.protobuf.json_format import MessageToDict
 
 from sglang.srt.disaggregation.utils import DisaggregationMode
-from sglang.srt.grpc import sglang_scheduler_pb2
-from sglang.srt.disaggregation.encode_receiver import MMReceiver
 from sglang.srt.managers.io_struct import (
     AbortReq,
     BatchEmbeddingOutput,
     BatchTokenIDOutput,
-    GenerateReqInput,
     GetLoadsReqInput,
     GetLoadsReqOutput,
     HealthCheckOutput,
@@ -38,14 +34,8 @@ from sglang.srt.observability.req_time_stats import (
     calibrate_time_diff,
     real_time,
 )
-from sglang.srt.sampling.sampling_params import SamplingParams as SGLSamplingParams
 from sglang.srt.server_args import PortArgs, ServerArgs
-from sglang.srt.utils import (
-    ImageData,
-    get_or_create_event_loop,
-    get_zmq_socket,
-    kill_process_tree,
-)
+from sglang.srt.utils import get_or_create_event_loop, get_zmq_socket, kill_process_tree
 from sglang.utils import get_exception_traceback
 
 logger = logging.getLogger(__name__)
@@ -228,12 +218,6 @@ class GrpcRequestManager:
 
         # Bootstrap server (passed from serve_grpc, not started here)
         self.bootstrap_server = bootstrap_server
-        self.mm_receiver: Optional[MMReceiver] = None
-        if (
-            self.server_args.language_only
-            and self.server_args.encoder_transfer_backend == "zmq_to_scheduler"
-        ):
-            self.mm_receiver = MMReceiver(self.server_args)
 
         # Communicators for request/response patterns with scheduler
         # Note: These must be initialized after send_to_scheduler socket is created
@@ -878,422 +862,6 @@ class GrpcRequestManager:
                     "request_id": getattr(obj, "rid", "unknown"),
                     "type": type(obj).__name__,
                 }
-            )
-
-    # Protocol conversion methods
-
-    def convert_generate_request(
-        self, grpc_req: sglang_scheduler_pb2.GenerateRequest
-    ) -> TokenizedGenerateReqInput:
-        """Convert gRPC GenerateRequest to internal format."""
-        if not grpc_req.HasField("tokenized"):
-            raise ValueError("Tokenized input must be provided")
-
-        input_text = grpc_req.tokenized.original_text
-        input_ids = list(grpc_req.tokenized.input_ids)
-
-        sampling_params = self._convert_sampling_params(grpc_req.sampling_params)
-        sampling_params.normalize(tokenizer=None)
-
-        bootstrap_host = None
-        bootstrap_port = None
-        bootstrap_room = None
-        if grpc_req.HasField("disaggregated_params"):
-            bootstrap_host = (
-                grpc_req.disaggregated_params.bootstrap_host
-                if grpc_req.disaggregated_params.bootstrap_host
-                else None
-            )
-            bootstrap_port = (
-                grpc_req.disaggregated_params.bootstrap_port
-                if grpc_req.disaggregated_params.bootstrap_port
-                else None
-            )
-            bootstrap_room = grpc_req.disaggregated_params.bootstrap_room
-
-        return TokenizedGenerateReqInput(
-            rid=grpc_req.request_id,
-            input_text=input_text,
-            input_ids=input_ids,
-            mm_inputs=None,
-            sampling_params=sampling_params,
-            return_logprob=grpc_req.return_logprob,
-            logprob_start_len=(
-                grpc_req.logprob_start_len
-                if grpc_req.logprob_start_len is not None
-                else -1
-            ),
-            top_logprobs_num=grpc_req.top_logprobs_num or 0,
-            stream=grpc_req.stream or False,
-            lora_id=grpc_req.lora_id if grpc_req.lora_id else None,
-            token_ids_logprob=(
-                list(grpc_req.token_ids_logprob) if grpc_req.token_ids_logprob else None
-            ),
-            bootstrap_host=bootstrap_host,
-            bootstrap_port=bootstrap_port,
-            bootstrap_room=bootstrap_room,
-        )
-
-    def _handle_epd_disaggregation_encode_request(
-        self,
-        grpc_req: sglang_scheduler_pb2.GenerateRequest,
-        tokenized_req: TokenizedGenerateReqInput,
-    ) -> None:
-        if (
-            not self.mm_receiver
-            or not self.server_args.language_only
-            or self.server_args.encoder_transfer_backend != "zmq_to_scheduler"
-        ):
-            return
-
-        image_urls = list(grpc_req.mm_inputs.image_urls)
-        if not image_urls:
-            return
-
-        encode_req = GenerateReqInput(
-            image_data=[ImageData(url=url) for url in image_urls],
-            rid=grpc_req.request_id,
-        )
-        self.mm_receiver.send_encode_request(encode_req)
-        tokenized_req.need_wait_for_image = bool(encode_req.need_wait_for_image)
-        tokenized_req.num_items_assigned = encode_req.num_items_assigned
-
-    def convert_embed_request(
-        self, grpc_req: sglang_scheduler_pb2.EmbedRequest
-    ) -> TokenizedEmbeddingReqInput:
-        """Convert gRPC EmbedRequest to internal format."""
-        if not grpc_req.HasField("tokenized"):
-            raise ValueError("Tokenized input must be provided")
-
-        input_text = grpc_req.tokenized.original_text
-        input_ids = list(grpc_req.tokenized.input_ids)
-
-        sampling_params = self._convert_sampling_params(grpc_req.sampling_params)
-        sampling_params.max_new_tokens = 0
-        sampling_params.normalize(tokenizer=None)
-
-        return TokenizedEmbeddingReqInput(
-            rid=grpc_req.request_id,
-            input_text=input_text,
-            input_ids=input_ids,
-            image_inputs={"mm_items": []},
-            token_type_ids=list(grpc_req.token_type_ids),
-            sampling_params=sampling_params,
-        )
-
-    def _convert_sampling_params(
-        self, grpc_params: sglang_scheduler_pb2.SamplingParams
-    ) -> SGLSamplingParams:
-        """Convert gRPC SamplingParams to internal format."""
-        regex = None
-        json_schema = None
-        ebnf_grammar = None
-        structural_tag = None
-
-        if grpc_params.HasField("regex"):
-            regex = grpc_params.regex
-        elif grpc_params.HasField("json_schema"):
-            json_schema = grpc_params.json_schema
-        elif grpc_params.HasField("ebnf_grammar"):
-            ebnf_grammar = grpc_params.ebnf_grammar
-        elif grpc_params.HasField("structural_tag"):
-            structural_tag = grpc_params.structural_tag
-
-        custom_params = (
-            MessageToDict(grpc_params.custom_params)
-            if grpc_params.HasField("custom_params")
-            else None
-        )
-        max_new_tokens = (
-            grpc_params.max_new_tokens
-            if grpc_params.HasField("max_new_tokens")
-            else None
-        )
-        stream_interval = (
-            grpc_params.stream_interval
-            if grpc_params.HasField("stream_interval")
-            else None
-        )
-        logit_bias = dict(grpc_params.logit_bias) if grpc_params.logit_bias else None
-        stop = list(grpc_params.stop) if grpc_params.stop else None
-        stop_token_ids = (
-            list(grpc_params.stop_token_ids) if grpc_params.stop_token_ids else None
-        )
-
-        return SGLSamplingParams(
-            temperature=grpc_params.temperature,
-            top_p=grpc_params.top_p,
-            top_k=grpc_params.top_k,
-            min_p=grpc_params.min_p,
-            frequency_penalty=grpc_params.frequency_penalty,
-            presence_penalty=grpc_params.presence_penalty,
-            repetition_penalty=grpc_params.repetition_penalty,
-            max_new_tokens=max_new_tokens,
-            min_new_tokens=grpc_params.min_new_tokens,
-            stop=stop,
-            stop_token_ids=stop_token_ids,
-            skip_special_tokens=grpc_params.skip_special_tokens,
-            spaces_between_special_tokens=grpc_params.spaces_between_special_tokens,
-            no_stop_trim=grpc_params.no_stop_trim,
-            regex=regex,
-            json_schema=json_schema,
-            ebnf=ebnf_grammar,
-            structural_tag=structural_tag,
-            n=grpc_params.n,
-            ignore_eos=grpc_params.ignore_eos,
-            stream_interval=stream_interval,
-            logit_bias=logit_bias,
-            custom_params=custom_params,
-        )
-
-    def create_generate_response(
-        self, request_id: str, output: Dict, is_final: bool = False
-    ) -> sglang_scheduler_pb2.GenerateResponse:
-        """Create gRPC GenerateResponse from internal output."""
-        if "error" in output:
-            return sglang_scheduler_pb2.GenerateResponse(
-                request_id=request_id,
-                error=sglang_scheduler_pb2.GenerateError(
-                    message=output["error"],
-                    http_status_code=(
-                        "500" if "abort" not in output else "499"
-                    ),
-                ),
-            )
-
-        if is_final or output.get("finished", False):
-            return self._create_completion_response(request_id, output)
-        else:
-            return self._create_chunk_response(request_id, output)
-
-    def create_embed_response(
-        self, request_id: str, result: Dict
-    ) -> sglang_scheduler_pb2.EmbedResponse:
-        """Create gRPC EmbedResponse from internal result."""
-        return sglang_scheduler_pb2.EmbedResponse(
-            request_id=request_id,
-            complete=sglang_scheduler_pb2.EmbedComplete(
-                embedding=result["embedding"],
-                prompt_tokens=result.get("prompt_tokens", 0),
-                cached_tokens=0,
-                embedding_dim=len(result["embedding"]),
-            ),
-        )
-
-    def _create_chunk_response(
-        self, request_id: str, output: Dict
-    ) -> sglang_scheduler_pb2.GenerateResponse:
-        """Create a streaming chunk response."""
-        meta_info = output.get("meta_info", {})
-
-        output_logprobs_proto = self._convert_output_logprobs_to_proto(
-            output.get("output_logprobs")
-        )
-        input_logprobs_proto = self._convert_input_logprobs_to_proto(
-            output.get("input_logprobs")
-        )
-
-        return sglang_scheduler_pb2.GenerateResponse(
-            request_id=request_id,
-            chunk=sglang_scheduler_pb2.GenerateStreamChunk(
-                token_ids=output.get("token_ids", []),
-                prompt_tokens=meta_info.get("prompt_tokens", 0),
-                completion_tokens=meta_info.get("completion_tokens", 0),
-                cached_tokens=meta_info.get("cached_tokens", 0),
-                output_logprobs=output_logprobs_proto,
-                input_logprobs=input_logprobs_proto,
-                index=output.get("index", 0),
-            ),
-        )
-
-    def _create_completion_response(
-        self, request_id: str, output: Dict
-    ) -> sglang_scheduler_pb2.GenerateResponse:
-        """Create a completion response."""
-        meta_info = output.get("meta_info", {})
-        finish_reason_data = meta_info.get("finish_reason")
-
-        finish_reason = "stop"
-        if finish_reason_data:
-            if isinstance(finish_reason_data, dict):
-                finish_reason_type = finish_reason_data.get("type")
-            else:
-                finish_reason_type = finish_reason_data
-
-            if finish_reason_type == "length":
-                finish_reason = "length"
-            elif finish_reason_type == "abort":
-                finish_reason = "abort"
-
-        matched_stop_kwargs = {}
-        if isinstance(finish_reason_data, dict) and "matched" in finish_reason_data:
-            matched = finish_reason_data["matched"]
-            if isinstance(matched, int):
-                matched_stop_kwargs["matched_token_id"] = matched
-            elif isinstance(matched, str):
-                matched_stop_kwargs["matched_stop_str"] = matched
-
-        output_logprobs_proto = self._convert_output_logprobs_to_proto(
-            output.get("output_logprobs")
-        )
-        input_logprobs_proto = self._convert_input_logprobs_to_proto(
-            output.get("input_logprobs")
-        )
-
-        return sglang_scheduler_pb2.GenerateResponse(
-            request_id=request_id,
-            complete=sglang_scheduler_pb2.GenerateComplete(
-                output_ids=output.get("token_ids", []),
-                finish_reason=finish_reason,
-                prompt_tokens=meta_info.get("prompt_tokens", 0),
-                completion_tokens=meta_info.get(
-                    "completion_tokens", len(output.get("token_ids", []))
-                ),
-                cached_tokens=meta_info.get("cached_tokens", 0),
-                output_logprobs=output_logprobs_proto,
-                input_logprobs=input_logprobs_proto,
-                index=output.get("index", 0),
-                **matched_stop_kwargs,
-            ),
-        )
-
-    def _convert_output_logprobs_to_proto(
-        self, logprobs_data: Dict
-    ) -> Optional[sglang_scheduler_pb2.OutputLogProbs]:
-        """Convert output logprobs dict to proto."""
-        if not logprobs_data:
-            return None
-
-        token_logprobs_val = logprobs_data.get("token_logprobs_val", [])
-        token_logprobs_idx = logprobs_data.get("token_logprobs_idx", [])
-        top_logprobs_val = logprobs_data.get("top_logprobs_val", [])
-        top_logprobs_idx = logprobs_data.get("top_logprobs_idx", [])
-
-        top_logprobs_proto = []
-        if top_logprobs_val and top_logprobs_idx:
-            for val_list, idx_list in zip(top_logprobs_val, top_logprobs_idx):
-                top_logprobs_proto.append(
-                    sglang_scheduler_pb2.TopLogProbs(
-                        values=val_list,
-                        token_ids=idx_list,
-                    )
-                )
-
-        return sglang_scheduler_pb2.OutputLogProbs(
-            token_logprobs=token_logprobs_val,
-            token_ids=token_logprobs_idx,
-            top_logprobs=top_logprobs_proto,
-        )
-
-    def _convert_input_logprobs_to_proto(
-        self, logprobs_data: Dict
-    ) -> Optional[sglang_scheduler_pb2.InputLogProbs]:
-        """Convert input logprobs dict to proto."""
-        if not logprobs_data:
-            return None
-
-        token_logprobs_val = logprobs_data.get("token_logprobs_val", [])
-        token_logprobs_idx = logprobs_data.get("token_logprobs_idx", [])
-        top_logprobs_val = logprobs_data.get("top_logprobs_val", [])
-        top_logprobs_idx = logprobs_data.get("top_logprobs_idx", [])
-
-        token_logprobs_wrapped = [
-            (
-                sglang_scheduler_pb2.InputTokenLogProb()
-                if x is None
-                else sglang_scheduler_pb2.InputTokenLogProb(value=x)
-            )
-            for x in token_logprobs_val
-        ]
-
-        top_logprobs_proto = []
-        if top_logprobs_val and top_logprobs_idx:
-            for val_list, idx_list in zip(top_logprobs_val, top_logprobs_idx):
-                top_logprobs_proto.append(
-                    sglang_scheduler_pb2.TopLogProbs(
-                        values=val_list,
-                        token_ids=idx_list,
-                    )
-                )
-
-        return sglang_scheduler_pb2.InputLogProbs(
-            token_logprobs=token_logprobs_wrapped,
-            token_ids=token_logprobs_idx,
-            top_logprobs=top_logprobs_proto,
-        )
-
-    # High-level request handlers
-
-    async def handle_generate_request(
-        self,
-        grpc_req: sglang_scheduler_pb2.GenerateRequest,
-        grpc_context: Optional[grpc.aio.ServicerContext] = None,
-    ) -> AsyncGenerator[sglang_scheduler_pb2.GenerateResponse, None]:
-        """Handle a gRPC generate request end-to-end."""
-        try:
-            tokenized_req = self.convert_generate_request(grpc_req)
-            self._handle_epd_disaggregation_encode_request(grpc_req, tokenized_req)
-
-            response_generator = self.generate_request(
-                obj=tokenized_req,
-                request_id=grpc_req.request_id,
-                grpc_context=grpc_context,
-            )
-
-            async for output in response_generator:
-                if isinstance(output, list):
-                    for batch_output in output:
-                        yield self.create_generate_response(
-                            grpc_req.request_id, batch_output, is_final=True
-                        )
-                else:
-                    yield self.create_generate_response(
-                        grpc_req.request_id,
-                        output,
-                        is_final=output.get("finished", False),
-                    )
-
-        except Exception as e:
-            logger.error(
-                f"Generate failed for request {grpc_req.request_id}: {e}\n"
-                f"{get_exception_traceback()}"
-            )
-            yield sglang_scheduler_pb2.GenerateResponse(
-                request_id=grpc_req.request_id,
-                error=sglang_scheduler_pb2.GenerateError(
-                    message=str(e),
-                    http_status_code="500",
-                    details=get_exception_traceback(),
-                ),
-            )
-
-    async def handle_embed_request(
-        self, grpc_req: sglang_scheduler_pb2.EmbedRequest
-    ) -> sglang_scheduler_pb2.EmbedResponse:
-        """Handle a gRPC embed request end-to-end."""
-        try:
-            tokenized_req = self.convert_embed_request(grpc_req)
-
-            future = await self.embedding_request(
-                obj=tokenized_req,
-                request_id=grpc_req.request_id,
-            )
-
-            result = await future
-            return self.create_embed_response(grpc_req.request_id, result)
-
-        except Exception as e:
-            logger.error(
-                f"Embed failed for request {grpc_req.request_id}: {e}\n"
-                f"{get_exception_traceback()}"
-            )
-            return sglang_scheduler_pb2.EmbedResponse(
-                request_id=grpc_req.request_id,
-                error=sglang_scheduler_pb2.EmbedError(
-                    message=str(e),
-                    code="INTERNAL_ERROR",
-                    details=get_exception_traceback(),
-                ),
             )
 
     async def shutdown(self):
